Enum quick_xml::de::DeError

pub enum DeError {
    Custom(String),
    InvalidXml(Error),
    InvalidInt(ParseIntError),
    InvalidFloat(ParseFloatError),
    InvalidBoolean(String),
    KeyNotRead,
    UnexpectedStart(Vec<u8>),
    UnexpectedEnd(Vec<u8>),
    UnexpectedEof,
    ExpectedStart,
    Unsupported(Cow<'static, str>),
    TooManyEvents(NonZeroUsize),
}

(De)serialization error

Variants

Custom(String)

Serde custom error

InvalidXml(Error)

Xml parsing error

InvalidInt(ParseIntError)

Cannot parse to integer

InvalidFloat(ParseFloatError)

Cannot parse to float

InvalidBoolean(String)

Cannot parse specified value to boolean

KeyNotRead

This error indicates an error in the Deserialize implementation when read a map or a struct: MapAccess::next_value[_seed] was called before MapAccess::next_key[_seed].

You should check your types, that implements corresponding trait.

UnexpectedStart(Vec<u8>)

Deserializer encounter a start tag with a specified name when it is not expecting. This happens when you try to deserialize a primitive value (numbers, strings, booleans) from an XML element.

UnexpectedEnd(Vec<u8>)

Deserializer encounter an end tag with a specified name when it is not expecting. Usually that should not be possible, because XML reader is not able to produce such stream of events that lead to this error.

If you get this error this likely indicates and error in the quick_xml. Please open an issue at https://github.com/tafia/quick-xml, provide your Rust code and XML input.

UnexpectedEof

The Reader produced Event::Eof when it is not expecting, for example, after producing Event::Start but before corresponding Event::End.

ExpectedStart

This error indicates that deserialize_struct was called, but there is no any XML element in the input. That means that you try to deserialize a struct not from an XML element.

Unsupported(Cow<'static, str>)

An attempt to deserialize to a type, that is not supported by the XML store at current position, for example, attempt to deserialize struct from attribute or attempt to deserialize binary data.

TooManyEvents(NonZeroUsize)

Too many events were skipped while deserializing a sequence, event limit exceeded. The limit was provided as an argument