quick_xml::de::key

Struct QNameDeserializer

pub struct QNameDeserializer<'i, 'd> {
    name: CowRef<'i, 'd, str>,
}

A deserializer for xml names of elements and attributes.

Used for deserializing values from:

attribute names (<... name="..." ...>)

element names (<name>...</name>)

Converts a name to an identifier string using the following rules:

if it is an attribute name, put @ in front of the identifier

if it is a namespace binding (xmlns or xmlns:xxx) put the decoded name to the identifier

if it is an attribute in the xml namespace, put the decoded name to the identifier

put the decoded local_name() of a name to the identifier

The final identifier looks like [@]local_name, or @xmlns, or @xmlns:binding or xml:attribute (where [] means optional element).

The deserializer also supports deserializing names as other primitive types:

numbers

booleans

unit (()) and unit structs

unit variants of the enumerations

Because serde does not define on which side type conversion should be performed, and because Deserialize implementation for that primitives in serde does not accept strings, the deserializer will perform conversion by itself.

The deserializer is able to deserialize unit and unit structs, but any name will be converted to the same unit instance. This is asymmetry with a serializer, which not able to serialize those types, because empty names are impossible in XML.

deserialize_any() returns the same result as deserialize_identifier().

Lifetimes

'i: lifetime of the data that the deserializer borrows from the parsed input

'd: lifetime of a deserializer that holds a buffer with content of events

Fields
name: CowRef<'i, 'd, str>
Implementations

impl<'i, 'd> QNameDeserializer<'i, 'd>

pub fn from_attr( name: QName<'d>, decoder: Decoder, key_buf: &'d mut String, ) -> Result<Self, DeError>

Creates deserializer from name of an attribute

pub fn from_elem(start: &'d BytesStart<'i>) -> Result<Self, DeError>

Creates deserializer from name of an element