quick_xml::de

Module map

Serde Deserializer module

Structs

ElementDeserializer 🔒

A deserializer for a single tag item of a mixed sequence of tags and text.

ElementMapAccess 🔒

A deserializer that extracts map-like structures from an XML. This deserializer represents a one XML tag:

MapValueDeserializer 🔒

A deserializer for a value of map or struct. That deserializer slightly differently processes events for a primitive types and sequences than a Deserializer.

MapValueSeqAccess 🔒

An accessor to sequence elements forming a value for struct field. Technically, this sequence is flattened out into structure and sequence elements are overlapped with other fields of a structure. Each call to Self::next_element_seed consumes a next sub-tree or consequent list of Text and CData events.

MapValueVariantAccess 🔒

Enums

TagFilter 🔒

A filter that determines, what tags should form a sequence.

ValueSource 🔒

Defines a source that should be used to deserialize a value in the next call to next_value_seed()

Functions

not_in 🔒

Check if tag start is included in the fields list. decoder is used to get a string representation of a tag.