quick_xml::de::map

Struct ElementDeserializer

struct ElementDeserializer<'de, 'd, R, E>
where
    R: XmlRead<'de>,
    E: EntityResolver,
{
    start: BytesStart<'de>,
    de: &'d mut Deserializer<'de, R, E>,
}

A deserializer for a single tag item of a mixed sequence of tags and text.

This deserializer are very similar to a MapValueDeserializer (when it processes the DeEvent::Start event). The only difference in the deserialize_seq method. This deserializer will perform deserialization from the textual content between start and end events, whereas the MapValueDeserializer will iterate over tags / text within it’s parent tag.

This deserializer processes items as following:

numbers are parsed from a text content between tags using FromStr. So, <int>123</int> can be deserialized into an u32;

booleans converted from a text content between tags according to the XML specification:

"true" and "1" converted to true;

"false" and "0" converted to false;

strings returned as a text content between tags;

characters also returned as strings. If string contain more than one character or empty, it is responsibility of a type to return an error;

Option are always deserialized as Some using the same deserializer, including <tag/> or <tag></tag>;

units (()) and unit structs consumes the whole element subtree;

newtype structs forwards deserialization to the inner type using SimpleTypeDeserializer;

sequences, tuples and tuple structs are deserialized using SimpleTypeDeserializer (this is the difference): text content between tags is passed to SimpleTypeDeserializer;

structs and maps are deserialized using new instance of ElementMapAccess;

enums:

the variant name is deserialized using QNameDeserializer from the element name;

the content is deserialized using the same deserializer:

unit variants: consuming a subtree and return ();

newtype variants forwards deserialization to the inner type using this deserializer;

tuple variants: call deserialize_tuple of this deserializer;

struct variants: call deserialize_struct of this deserializer.

Fields
start: BytesStart<'de>de: &'d mut Deserializer<'de, R, E>
Implementations

impl<'de, 'd, R, E> ElementDeserializer<'de, 'd, R, E>
where R: XmlRead<'de>, E: EntityResolver,

fn read_string(&mut self) -> Result<Cow<'de, str>, DeError>

Returns a next string as concatenated content of consequent Text and CData events, used inside deserialize_primitives!().