quick_xml::de::map

Struct ElementMapAccess

pub(crate) struct ElementMapAccess<'de, 'd, R, E>
where
    R: XmlRead<'de>,
    E: EntityResolver,
{
    start: BytesStart<'de>,
    de: &'d mut Deserializer<'de, R, E>,
    iter: IterState,
    source: ValueSource,
    fields: &'static [&'static str],
    has_value_field: bool,
    has_text_field: bool,
}

A deserializer that extracts map-like structures from an XML. This deserializer represents a one XML tag:

<tag>...</tag>

Name of this tag is stored in a Self::start property.

Lifetimes

'de lifetime represents a buffer, from which deserialized values can borrow their data. Depending on the underlying reader, there can be an internal buffer of deserializer (i.e. deserializer itself) or an input (in that case it is possible to approach zero-copy deserialization).

'd lifetime represents a parent deserializer, which could own the data buffer.

Fields
start: BytesStart<'de>

Tag – owner of attributes

de: &'d mut Deserializer<'de, R, E>iter: IterState

State of the iterator over attributes. Contains the next position in the inner start slice, from which next attribute should be parsed.

source: ValueSource

Current state of the accessor that determines what next call to API methods should return.

fields: &'static [&'static str]

List of field names of the struct. It is empty for maps

has_value_field: bool

If true, then the deserialized struct has a field with a special name: VALUE_KEY. That field should be deserialized from the whole content of an XML node, including tag name:

<tag>value for VALUE_KEY field<tag>

has_text_field: bool

If true, then the deserialized struct has a field with a special name: TEXT_KEY.

Implementations

impl<'de, 'd, R, E> ElementMapAccess<'de, 'd, R, E>
where R: XmlRead<'de>, E: EntityResolver,

pub fn new( de: &'d mut Deserializer<'de, R, E>, start: BytesStart<'de>, fields: &'static [&'static str], ) -> Self

Create a new ElementMapAccess

fn should_skip_subtree(&self, start: &BytesStart<'_>) -> bool

Determines if subtree started with the specified event shoould be skipped.

Used to map elements with xsi:nil attribute set to true to None in optional contexts.

We need to handle two attributes:

on parent element: <map xsi:nil="true"><foo/></map>

on this element: <map><foo xsi:nil="true"/></map>

We check parent element too because xsi:nil affects only nested elements of the tag where it is defined. We can map structure with fields mapped to attributes to the <map> element and set to None all its optional elements.

fn skip_whitespaces(&mut self) -> Result<(), DeError>

Skips whitespaces when they are not preserved