quick_xml::de

Module simple_type

Contains Serde Deserializer for XML simple types as defined in the XML Schema.

Macros

deserialize_num 🔒

deserialize_primitive 🔒

unsupported 🔒

Structs

AtomicDeserializer 🔒

A deserializer that handles ordinary simple type definition with {variety} = atomic, or an ordinary simple type definition with {variety} = union whose basic members are all atomic.

ListIter 🔒

Iterator over string sub-slices delimited by one or several spaces. Contains decoded value of the simpleType. Iteration ends when list contains None.

SimpleTypeDeserializer

A deserializer for an xml probably escaped and encoded value of XSD simple types. This deserializer will borrow from the input as much as possible.

UnitOnly

Deserializer of variant data, that supports only unit variants. Attempt to deserialize newtype will provide UnitDeserializer. Attempt to deserialize tuple or struct variant will result to call of Visitor::visit_unit.

Enums

Content 🔒

A version of Cow that can borrow from two different buffers, one of them is a deserializer input, and conceptually contains only part of owned data.