quick_xml::de

Struct Deserializer

pub struct Deserializer<'de, R, E: EntityResolver = PredefinedEntityResolver>
where
    R: XmlRead<'de>,
{ /* private fields */ }

A structure that deserializes XML into Rust values.

Implementations

impl<'de, R, E> Deserializer<'de, R, E>
where R: XmlRead<'de>, E: EntityResolver,

pub fn is_empty(&self) -> bool

Returns true if all events was consumed.

pub const fn get_ref(&self) -> &R

Returns the underlying XML reader.

use serde::Deserialize;
use quick_xml::de::Deserializer;
use quick_xml::NsReader;

#[derive(Deserialize)]
struct SomeStruct {
    field1: String,
    field2: String,
}

// Try to deserialize from broken XML
let mut de = Deserializer::from_str(
    "<SomeStruct><field1><field2></SomeStruct>"
//   0                           ^= 28        ^= 41
);

let err = SomeStruct::deserialize(&mut de);
assert!(err.is_err());

let reader: &NsReader<_> = de.get_ref().get_ref();

assert_eq!(reader.error_position(), 28);
assert_eq!(reader.buffer_position(), 41);

impl<'de> Deserializer<'de, SliceReader<'de>>

pub fn from_str(source: &'de str) -> Self

Create a new deserializer that will borrow data from the specified string.

Deserializer created with this method will not resolve custom entities.

pub fn borrowing(reader: NsReader<&'de [u8]>) -> Self

Create a new deserializer that will borrow data from the specified preconfigured reader.

Deserializer created with this method will not resolve custom entities.

Note, that config option Config::expand_empty_elements will be set to true.

Example

#[derive(Deserialize, PartialEq, Debug)]
struct Object<'a> {
    tag: &'a str,
}

let mut reader = NsReader::from_str("<xml><tag>    test    </tag></xml>");

let mut de = Deserializer::borrowing(reader.clone());
let obj = Object::deserialize(&mut de).unwrap();
assert_eq!(obj, Object { tag: "    test    " });

reader.config_mut().trim_text(true);

let mut de = Deserializer::borrowing(reader);
let obj = Object::deserialize(&mut de).unwrap();
assert_eq!(obj, Object { tag: "test" });

impl<'de, E> Deserializer<'de, SliceReader<'de>, E>
where E: EntityResolver,

pub fn from_str_with_resolver(source: &'de str, entity_resolver: E) -> Self

Create a new deserializer that will borrow data from the specified string and use the specified entity resolver.

pub fn borrowing_with_resolver( reader: NsReader<&'de [u8]>, entity_resolver: E, ) -> Self

Create a new deserializer that will borrow data from the specified preconfigured reader and use the specified entity resolver.

Note, that config option Config::expand_empty_elements will be set to true.

impl<'de, R> Deserializer<'de, IoReader<R>>
where R: BufRead,

pub fn from_reader(reader: R) -> Self

Create a new deserializer that will copy data from the specified reader into internal buffer.

If you already have a string use Self::from_str instead, because it will borrow instead of copy. If you have &[u8] which is known to represent UTF-8, you can decode it first before using from_str.

Deserializer created with this method will not resolve custom entities.

pub fn buffering(reader: NsReader<R>) -> Self

Create a new deserializer that will copy data from the specified preconfigured reader into internal buffer.

Deserializer created with this method will not resolve custom entities.

Note, that config option Config::expand_empty_elements will be set to true.

Example

#[derive(Deserialize, PartialEq, Debug)]
struct Object {
    tag: String,
}

let mut reader = NsReader::from_str("<xml><tag>    test    </tag></xml>");

let mut de = Deserializer::buffering(reader.clone());
let obj = Object::deserialize(&mut de).unwrap();
assert_eq!(obj, Object { tag: "    test    ".to_string() });

reader.config_mut().trim_text(true);

let mut de = Deserializer::buffering(reader);
let obj = Object::deserialize(&mut de).unwrap();
assert_eq!(obj, Object { tag: "test".to_string() });

impl<'de, R, E> Deserializer<'de, IoReader<R>, E>
where R: BufRead, E: EntityResolver,

pub fn with_resolver(reader: R, entity_resolver: E) -> Self

Create a new deserializer that will copy data from the specified reader into internal buffer and use the specified entity resolver.

If you already have a string use Self::from_str instead, because it will borrow instead of copy. If you have &[u8] which is known to represent UTF-8, you can decode it first before using from_str.

pub fn buffering_with_resolver(reader: NsReader<R>, entity_resolver: E) -> Self

Create new deserializer that will copy data from the specified preconfigured reader into internal buffer and use the specified entity resolver.

Note, that config option Config::expand_empty_elements will be set to true.