quick_xml::de

Struct SimpleTypeDeserializer

pub struct SimpleTypeDeserializer<'de, 'a> { /* private fields */ }

A deserializer for an xml probably escaped and encoded value of XSD simple types. This deserializer will borrow from the input as much as possible.

deserialize_any() returns the whole string that deserializer contains.

Escaping the value is actually not always necessary, for instance when converting to a float, we don’t expect any escapable character anyway. In that cases deserializer skips unescaping step.

Used for deserialize values from:

attribute values (<... ...="value" ...>)

mixed text / CDATA content (<...>text<![CDATA[cdata]]></...>)

This deserializer processes items as following:

numbers are parsed from a text content using FromStr; in case of error Visitor::visit_borrowed_str, Visitor::visit_str, or Visitor::visit_string is called; it is responsibility of the type to return an error if it does not able to process passed data;

booleans converted from the text according to the XML specification:

"true" and "1" converted to true;

"false" and "0" converted to false;

everything else calls Visitor::visit_borrowed_str, Visitor::visit_str, or Visitor::visit_string; it is responsibility of the type to return an error if it does not able to process passed data;

strings returned as is;

characters also returned as strings. If string contain more than one character or empty, it is responsibility of a type to return an error;

Option always deserialized as Some using the same deserializer. If attribute or text content is missed, then the deserializer even wouldn’t be used, so if it is used, then the value should be;

units (()) and unit structs always deserialized successfully, the content is ignored;

newtype structs forwards deserialization to the inner type using the same deserializer;

sequences, tuples and tuple structs are deserialized as xs:lists. Only sequences of primitive types is possible to deserialize this way and they should be delimited by a space ( , \t, \r, or \n);

structs and maps delegates to Self::deserialize_str which calls Visitor::visit_borrowed_str or Visitor::visit_string; it is responsibility of the type to return an error if it does not able to process passed data;

enums:

the variant name is deserialized using the same deserializer;

the content is deserialized using the deserializer that always returns unit (()):

unit variants: just return ();

newtype variants: deserialize from UnitDeserializer;

tuple and struct variants: call Visitor::visit_unit;

identifiers are deserialized as strings.

Implementations

impl<'de, 'a> SimpleTypeDeserializer<'de, 'a>

pub fn from_text(text: Cow<'de, str>) -> Self

Creates a deserializer from a value, that possible borrowed from input.

It is assumed that text does not have entities.

pub fn from_text_content(value: Text<'de>) -> Self

Creates a deserializer from an XML text node, that possible borrowed from input.

It is assumed that text does not have entities.

This constructor used internally to deserialize from text nodes.