quick_xml::de::text

Struct TextDeserializer

pub struct TextDeserializer<'de>(pub Text<'de>);

A deserializer for a single text node of a mixed sequence of tags and text.

This deserializer are very similar to a MapValueDeserializer (when it processes the DeEvent::Text event). The only difference in the deserialize_seq method. This deserializer will perform deserialization from a textual content, whereas the MapValueDeserializer will iterate over tags / text within it’s parent tag.

This deserializer processes items as following:

numbers are parsed from a text content using FromStr; in case of error Visitor::visit_borrowed_str, Visitor::visit_str, or Visitor::visit_string is called; it is responsibility of the type to return an error if it does not able to process passed data;

booleans converted from the text according to the XML specification:

"true" and "1" converted to true;

"false" and "0" converted to false;

everything else calls Visitor::visit_borrowed_str, Visitor::visit_str, or Visitor::visit_string; it is responsibility of the type to return an error if it does not able to process passed data;

strings returned as is;

characters also returned as strings. If string contain more than one character or empty, it is responsibility of a type to return an error;

Option:

empty text is deserialized as None;

everything else is deserialized as Some using the same deserializer;

units (()) and unit structs always deserialized successfully, the content is ignored;

newtype structs forwards deserialization to the inner type using the same deserializer;

sequences, tuples and tuple structs are deserialized using SimpleTypeDeserializer (this is the difference): text content passed to the deserializer directly;

structs and maps calls Visitor::visit_borrowed_str or Visitor::visit_string, it is responsibility of the type to return an error if it do not able to process this data;

enums:

the variant name is deserialized as $text;

the content is deserialized using the same deserializer:

unit variants: just return ();

newtype variants forwards deserialization to the inner type using the same deserializer;

tuple and struct variants are deserialized using SimpleTypeDeserializer.

Tuple Fields
0: Text<'de>
Implementations

impl<'de> TextDeserializer<'de>

fn read_string(self) -> Result<Cow<'de, str>, DeError>

Returns a next string as concatenated content of consequent Text and CData events, used inside deserialize_primitives!().