quick_xml

Module errors

Error management module

Enums

Error

The error type used by this crate.

IllFormedError

An error returned if parsed document is not well-formed, for example, an opened tag is not closed before end of input.

SyntaxError

An error returned if parsed document does not correspond to the XML grammar, for example, a tag opened by < not closed with >. This error does not represent invalid XML constructs, for example, tags <> and </> a well-formed from syntax point-of-view.

Type Aliases

Result

A specialized Result type where the error is hard-wired to Error.