quick_xml::errors::serialize

Enum DeError

pub enum DeError {
    Custom(String),
    InvalidXml(Error),
    InvalidInt(ParseIntError),
    InvalidFloat(ParseFloatError),
    InvalidBoolean(String),
    KeyNotRead,
    UnexpectedStart(Vec<u8>),
    UnexpectedEof,
    Unsupported(Cow<'static, str>),
    TooManyEvents(NonZeroUsize),
}

(De)serialization error

Variants

Custom(String)

Serde custom error

InvalidXml(Error)

Xml parsing error

InvalidInt(ParseIntError)

Cannot parse to integer

InvalidFloat(ParseFloatError)

Cannot parse to float

InvalidBoolean(String)

Cannot parse specified value to boolean

KeyNotRead

This error indicates an error in the Deserialize implementation when read a map or a struct: MapAccess::next_value[_seed] was called before MapAccess::next_key[_seed].

You should check your types, that implements corresponding trait.

UnexpectedStart(Vec<u8>)

Deserializer encounter a start tag with a specified name when it is not expecting. This happens when you try to deserialize a primitive value (numbers, strings, booleans) from an XML element.

UnexpectedEof

The Reader produced Event::Eof when it is not expecting, for example, after producing Event::Start but before corresponding Event::End.

Unsupported(Cow<'static, str>)

An attempt to deserialize to a type, that is not supported by the XML store at current position, for example, attempt to deserialize struct from attribute or attempt to deserialize binary data.

Serialized type cannot be represented in an XML due to violation of the XML rules in the final XML document. For example, attempt to serialize a HashMap<{integer}, ...> would cause this error because XML name cannot start from a digit or a hyphen (minus sign). The same result would occur if map key is a complex type that cannot be serialized as a primitive type (i.e. string, char, bool, unit struct or unit variant).

TooManyEvents(NonZeroUsize)

Too many events were skipped while deserializing a sequence, event limit exceeded. The limit was provided as an argument