quick_xml::escape

Function normalize_attr_step

fn normalize_attr_step<'entity, C, E, F>(
    normalized: &mut String,
    iter: &mut Iter<'_, u8>,
    input: &str,
    last_pos: usize,
    index: usize,
    depth: usize,
    is_normalization_char: &C,
    normalize_eol_step: &E,
    resolve_entity: &mut F,
) -> Result<usize, EscapeError>
where
    C: Fn(&u8) -> bool,
    E: Fn(&mut String, &str, usize, char) -> usize,
    F: FnMut(&str) -> Option<&'entity str>,

Performs one step of the normalization algorithm (but with recursive part):

For a character reference, append the referenced character to the normalized value.

For an entity reference, recursively apply this algorithm to the replacement text of the entity.

For a white space character (#x20, #xD, #xA, #x9), append a space character (#x20) to the normalized value.

For another character, append the character to the normalized value.

Because according to the specification, XML parser should parse line-of-end normalized input, but quick-xml does not do that, this function also performs normalization of EOL characters. That should be done before expanding entities and character references, so cannot be processed later.

This function could be used also just to normalize line ends if the iterator won’t be stop on & characters.

Parameters

normalized: Output of the algorithm. Normalized value will be placed here

iter: Iterator over bytes of input

input: Original non-normalized value

last_pos: Index of the last byte in input that was processed

index: Index of the byte in input that should be processed now

seen_cr: \r\n and \r\x85 sequences should be normalized into one space so this parameter tracks if we seen the \r before processing the current byte

depth: Current recursion depth. Too deep recursion will interrupt the algorithm

is_normalization_char: a function to check if byte is the start byte of character that should be normalized (UTF-8 encoding is assumed)

normalize_eol_step: a function that performs EOL normalization of a character

resolve_entity: Resolver of entities. Returns None for unknown entities

Lifetimes

'entity: lifetime of all entities that is returned by the entity resolution routine