quick_xml::escape

Function normalize_xml10_eol_step

fn normalize_xml10_eol_step(
    normalized: &mut String,
    input: &[u8],
    index: usize,
    ch: char,
) -> usize

The text in HTML normally is not normalized in any way; normalization is performed only in limited contexts and only for \r\n and \r.

Parameters

normalized: the string with the result of normalization

input: UTF-8 bytes of the string to be normalized

index: a byte index into input of character which is processed right now. It always points to the first byte of character in UTF-8 encoding

ch: a character that should be put to the string instead of newline sequence