quick_xml::escape

Function normalize_xml11_eol_step

fn normalize_xml11_eol_step(
    normalized: &mut String,
    text: &str,
    index: usize,
    ch: char,
) -> usize

All line breaks MUST have been normalized on input to #xA as described in 2.11 End-of-Line Handling, so the rest of this algorithm operates on text normalized in this way.

To simplify the tasks of applications, the XML processor MUST behave as if it normalized all line breaks in external parsed entities (including the document entity) on input, before parsing, by translating all of the following to a single #xA character (which attribute normalization routine will replace by #x20 character):

the two-character sequence #xD #xA

the two-character sequence #xD #x85

the single character #x85

the single character #x2028

any #xD character that is not immediately followed by #xA or #x85.

The characters #x85 and #x2028 cannot be reliably recognized and translated until an entity’s encoding declaration (if present) has been read. Therefore, it is a fatal error to use them within the XML declaration or text declaration.

Note, that this function cannot be used to normalize HTML values. The text in HTML normally is not normalized in any way; normalization is performed only in limited contexts and only for \r\n and \r.

Parameters

normalized: the string with the result of normalization

input: UTF-8 bytes of the string to be normalized

index: a byte index into input of character which is processed right now. It always points to the first byte of character in UTF-8 encoding

ch: a character that should be put to the string instead of newline sequence

Returns the index of next unprocessed byte in the input.