Struct quick_xml::events::BytesStart

pub struct BytesStart<'a> { /* fields omitted */ }

Opening tag data (Event::Start), with optional attributes.

<name attr="value">.

The name can be accessed using the name, local_name or unescaped methods. An iterator over the attributes is returned by the attributes method.

Methods

impl<'a> BytesStart<'a>

pub fn borrowed(content: &'a [u8], name_len: usize) -> Self

Creates a new BytesStart from the given content (name + attributes).

Warning

&content[..name_len] is not checked to be a valid name

pub fn borrowed_name(name: &'a [u8]) -> BytesStart<'a>

Creates a new BytesStart from the given name.

Warning

&content is not checked to be a valid name

pub fn owned<C: Into<Vec<u8>>>(
content: C,
name_len: usize
) -> BytesStart<'static>

Creates a new BytesStart from the given content (name + attributes)

Owns its contents.

pub fn owned_name<C: Into<Vec<u8>>>(name: C) -> BytesStart<'static>

Creates a new BytesStart from the given name

Owns its contents.

pub fn into_owned(self) -> BytesStart<'static>

Converts the event into an owned event.

pub fn to_owned(&self) -> BytesStart<'static>

Converts the event into an owned event without taking ownershiph of Event

pub fn with_attributes<'b, I>(self, attributes: I) -> Self where
I: IntoIterator,
I::Item: Into<Attribute<'b>>,

Consumes self and yield a new BytesStart with additional attributes from an iterator.

The yielded items must be convertible to Attribute using Into.

pub fn name(&self) -> &[u8]

Gets the undecoded raw tag name as a &[u8].

pub fn local_name(&self) -> &[u8]

Gets the undecoded raw local tag name (excluding namespace) as a &[u8].

All content up to and including the first : character is removed from the tag name.

pub fn unescaped(&self) -> Result<Cow<[u8]>>

Gets the unescaped tag name.

XML escape sequences like "&lt;" will be replaced by their unescaped characters like "<".

Important traits for Attributes<'a>
impl<'a> Iterator for Attributes<'a> type Item = Result<Attribute<'a>>;

pub fn attributes(&self) -> Attributes

Returns an iterator over the attributes of this tag.

Important traits for Attributes<'a>
impl<'a> Iterator for Attributes<'a> type Item = Result<Attribute<'a>>;

pub fn html_attributes(&self) -> Attributes

Returns an iterator over the HTML-like attributes of this tag (no mandatory quotes or =).

pub fn extend_attributes<'b, I>(&mut self, attributes: I) -> &mut BytesStart<'a> where
I: IntoIterator,
I::Item: Into<Attribute<'b>>,

Add additional attributes to this tag using an iterator.

The yielded items must be convertible to Attribute using Into.

pub fn unescape_and_decode<B: BufRead>(
&self,
reader: &Reader<B>
) -> Result<String>

Returns the unescaped and decoded string value.

This allocates a String in all cases. For performance reasons it might be a better idea to instead use one of:

unescaped(), as it doesn't allocate when no escape sequences are used.

Reader::decode(), as it only allocates when the decoding can't be performed otherwise.

pub fn push_attribute<'b, A: Into<Attribute<'b>>>(&mut self, attr: A)

Adds an attribute to this element.

pub fn set_name(&mut self, name: &[u8]) -> &mut BytesStart<'a>

Edit the name of the BytesStart in-place

Warning

name is not checked to be a valid name

pub fn clear_attributes(&mut self) -> &mut BytesStart<'a>

Remove all attributes from the ByteStart