quick_xml

Module name

Module for handling names according to the W3C Namespaces in XML 1.1 (Second Edition) specification

Structs

LocalName

A local (unqualified) name of an element or an attribute, i.e. a name without prefix.

Namespace

A namespace name that is declared in a xmlns[:prefix]="namespace name".

NamespaceBindingsIter

Iterator on the current declared namespace bindings. Returns pairs of the (prefix, namespace).

NamespaceBindingsOfLevelIter

Iterator on the declared namespace bindings on specified level. Returns pairs of the (prefix, namespace).

NamespaceResolver

A storage for currently defined namespace bindings, which is used to resolve prefixes into namespaces.

Prefix

A namespace prefix part of the qualified name of an element tag or an attribute: a prefix in <prefix:local-element-name> or prefix:local-attribute-name="attribute value".

QName

A qualified name of an element or an attribute, including an optional namespace prefix and a local name.

Enums

NamespaceError

Some namespace was invalid

PrefixDeclaration

A namespace prefix declaration, xmlns or xmlns:<name>, as defined in XML Schema specification

ResolveResult

Result of prefix resolution which creates by NamespaceResolver::resolve, NsReader::read_resolved_event and NsReader::read_resolved_event_into methods.

Constants

DEFAULT_MAX_DECLARATIONS_PER_ELEMENT

Default limit on the number of xmlns / xmlns:* declarations NamespaceResolver::push will accept on a single start tag.