quick_xml
Module parser

Contains low-level parsers of different XML pieces.

Structs

PiParser

A parser that search a ?> sequence in the slice.

Enums

ElementParser

A parser that search a > symbol in the slice outside of quoted regions.

Traits

Parser

Used to decouple reading of data from data source and parsing XML structure from it. This is a state preserved between getting chunks of bytes from the reader.