quick_xml::se

Module content

Contains serializer for content of an XML element

Macros

write_primitive 🔒

Structs

ContentSerializer

A serializer used to serialize content of an element. It does not write surrounding tags. Unlike the ElementSerializer, this serializer serializes enums using variant names as tag names, i. e. as <variant>...</variant>.

Seq

Helper struct which remembers the classification of the last serialized element and reports it when the sequence ends