quick_xml::se::content

Struct ContentSerializer

pub struct ContentSerializer<'w, 'i, W: Write> {
    pub writer: &'w mut W,
    pub level: QuoteLevel,
    pub(super) indent: Indent<'i>,
    pub write_indent: bool,
    pub text_format: TextFormat,
    pub allow_primitive: bool,
    pub expand_empty_elements: bool,
}

A serializer used to serialize content of an element. It does not write surrounding tags. Unlike the ElementSerializer, this serializer serializes enums using variant names as tag names, i. e. as <variant>...</variant>.

Returns the classification of the last written type.

This serializer does the following:

numbers converted to a decimal representation and serialized as naked strings;

booleans serialized ether as "true" or "false";

strings and characters are serialized as naked strings;

None does not write anything;

Some and newtypes are serialized as an inner type using the same serializer;

units (()) and unit structs does not write anything;

sequences, tuples and tuple structs are serialized without delimiters. [1, 2, 3] would be serialized as 123 (if not using indent);

structs and maps are not supported (SeError::Unsupported is returned);

enums:

unit variants are serialized as self-closed <variant/>;

newtype variants are serialized as inner value wrapped in <variant>...</variant>;

tuple variants are serialized as sequences where each element is wrapped in <variant>...</variant>;

struct variants are serialized as a sequence of fields wrapped in <variant>...</variant>. Each field is serialized recursively using either ElementSerializer, ContentSerializer ($value fields), or SimpleTypeSerializer ($text fields). In particular, the empty struct is serialized as <variant/>;

Usage of empty tags depends on the Self::expand_empty_elements setting.

The difference between this serializer and SimpleTypeSerializer is in how sequences and maps are serialized. Unlike SimpleTypeSerializer it supports any types in sequences and serializes them as list of elements, but that has drawbacks. Sequence of primitives would be serialized without delimiters and it will be impossible to distinguish between them. Even worse, when serializing with indent, sequence of strings become one big string with additional content and it would be impossible to distinguish between content of the original strings and inserted indent characters.

Fields
writer: &'w mut Wlevel: QuoteLevel

Defines which XML characters need to be escaped in text content

indent: Indent<'i>

Current indentation level. Note, that Indent::None means that there is no indentation at all, but write_indent == false means only, that indent writing is disabled in this instantiation of ContentSerializer, but child serializers should have access to the actual state of indentation.

write_indent: bool

If true, then current indent will be written before writing the content, but only if content is not empty. This flag is reset after writing indent.

text_format: TextFormat

Defines how text content should be serialized (as escaped text or CDATA)

allow_primitive: bool

If true, then primitive types that serializes to a text content without surrounding tag will be allowed, otherwise the SeError::Unsupported will be returned.

This method protects from the situation when two consequent values serialized as a text that makes it impossible to distinguish between them during deserialization. Instead of ambiguous serialization the error is returned.

expand_empty_elements: bool
Implementations

impl<'w, 'i, W: Write> ContentSerializer<'w, 'i, W>

pub fn into_simple_type_serializer_impl(self) -> SimpleTypeSerializer<&'w mut W>

Turns this serializer into serializer of a text content

pub fn into_simple_type_serializer( self, ) -> Result<SimpleTypeSerializer<&'w mut W>, SeError>

Turns this serializer into serializer of a text content if that is allowed, otherwise error is returned

pub fn new_seq_element_serializer( &mut self, allow_primitive: bool, ) -> ContentSerializer<'_, '_, W>

Creates new serializer that shares state with this serializer and writes to the same underlying writer

pub(super) fn write_empty( self, name: XmlName<'_>, ) -> Result<WriteResult, SeError>

Writes name as self-closed tag

pub(super) fn write_wrapped<S>( self, name: XmlName<'_>, serialize: S, ) -> Result<WriteResult, SeError>
where S: for<'a> FnOnce(SimpleTypeSerializer<&'a mut W>) -> Result<&'a mut W, SeError>,

Writes simple type content between name tags

pub(super) fn write_indent(&mut self) -> Result<(), SeError>