quick_xml::se

Module element

Contains serializer for an XML element

Macros

write_primitive 🔒

Writes simple type content between ElementSerializer::key tags.

Structs

ElementSerializer

A serializer used to serialize element with specified name. Unlike the ContentSerializer, this serializer never uses variant names of enum variants, and because of that it is unable to serialize any enum values, except unit variants.

Map

Struct

A serializer for struct variants, which serializes the struct contents inside of wrapping tags (<${tag}>...</${tag}>).

Enums

Tuple

A serializer for tuple variants. Tuples can be serialized in two modes: