quick_xml::se::element

Struct ElementSerializer

pub struct ElementSerializer<'w, 'k, W: Write> {
    pub ser: ContentSerializer<'w, 'k, W>,
    pub(super) key: XmlName<'k>,
}

A serializer used to serialize element with specified name. Unlike the ContentSerializer, this serializer never uses variant names of enum variants, and because of that it is unable to serialize any enum values, except unit variants.

Returns the classification of the last written type.

This serializer is used for an ordinary fields in structs, which are not special fields named $text (TEXT_KEY) or $value (VALUE_KEY). $text field should be serialized using SimpleTypeSerializer and $value field should be serialized using ContentSerializer.

This serializer does the following:

numbers converted to a decimal representation and serialized as <key>value</key>;

booleans serialized ether as <key>true</key> or <key>false</key>;

strings and characters are serialized as <key>value</key>. In particular, an empty string is serialized as <key/>;

None is serialized as <key/>;

Some and newtypes are serialized as an inner type using the same serializer;

units (()) and unit structs are serialized as <key/>;

sequences, tuples and tuple structs are serialized as repeated <key> tag. In particular, empty sequence is serialized to nothing;

structs are serialized as a sequence of fields wrapped in a <key> tag. Each field is serialized recursively using either ElementSerializer, ContentSerializer ($value fields), or SimpleTypeSerializer ($text fields). In particular, the empty struct is serialized as <key/>;

maps are serialized as a sequence of entries wrapped in a <key> tag. If key is serialized to a special name, the same rules as for struct fields are applied. In particular, the empty map is serialized as <key/>;

enums:

unit variants are serialized as <key>variant</key>;

other variants are not supported (SeError::Unsupported is returned);

Usage of empty tags depends on the ContentSerializer::expand_empty_elements setting.

Fields
ser: ContentSerializer<'w, 'k, W>

The inner serializer that contains the settings and mostly do the actual work

key: XmlName<'k>

Tag name used to wrap serialized types except enum variants which uses the variant name