quick_xml::se

Module simple_type

Contains Serde Serializer for XML simple types as defined in the XML Schema.

Macros

write_atomic 🔒

Structs

AtomicSerializer

A serializer that handles ordinary simple type definition with {variety} = atomic, or an ordinary simple type definition with {variety} = union whose basic members are all atomic.

SimpleSeq

Serializer for a sequence of atomic values delimited by space

SimpleTypeSerializer

A serializer for a values representing XSD simple types, which used in:

Enums

QuoteTarget

Functions

escape_into 🔒

escape_item 🔒

Escapes atomic value that could be part of a xs:list. All whitespace characters additionally escaped

escape_list 🔒

Escapes XSD simple type value