Struct quick_xml::writer::ElementWriter

pub struct ElementWriter<'a, W: Write> { /* private fields */ }

A struct to write an element. Contains methods to add attributes and inner elements to the element

Implementations

impl<'a, W: Write> ElementWriter<'a, W>

pub fn with_attribute<'b, I>(self, attr: I) -> Selfwhere
I: Into<Attribute<'b>>,

Adds an attribute to this element.

pub fn with_attributes<'b, I>(self, attributes: I) -> Selfwhere
I: IntoIterator,
I::Item: Into<Attribute<'b>>,

Add additional attributes to this element using an iterator.

The yielded items must be convertible to Attribute using Into.

pub fn write_text_content(self, text: BytesText<'_>) -> Result<&'a mut Writer<W>>

Write some text inside the current element.

pub fn write_cdata_content(
self,
text: BytesCData<'_>
) -> Result<&'a mut Writer<W>>

Write a CData event <![CDATA[...]]> inside the current element.

pub fn write_pi_content(self, text: BytesText<'_>) -> Result<&'a mut Writer<W>>

Write a processing instruction <?...?> inside the current element.

pub fn write_empty(self) -> Result<&'a mut Writer<W>>

Write an empty (self-closing) tag.

pub fn write_inner_content<F>(self, closure: F) -> Result<&'a mut Writer<W>>where
F: Fn(&mut Writer<W>) -> Result<()>,

Create a new scope for writing XML inside the current element.