Module quick_xml::de

Serde Deserializer module

Structs

Deserializer

A structure that deserializes XML into Rust values.

IoReader

XML input source that reads from a std::io input stream.

SliceReader

XML input source that reads from a slice of bytes and can borrow from it.

Enums

DeError

(De)serialization error

DeEvent

Simplified event which contains only these variants that used by deserializer

Traits

XmlRead

Trait used by the deserializer for iterating over input. This is manually “specialized” for iterating over &[u8].

Functions

from_reader

Deserialize from a reader. This method will do internal copies of data readed from reader. If you want have a &str input and want to borrow as much as possible, use from_str.

from_str

Deserialize an instance of type T from a string of XML text.