Struct quick_xml::de::Deserializer

pub struct Deserializer<'de, R>where
    R: XmlRead<'de>,{ /* private fields */ }

A structure that deserializes XML into Rust values.

Implementations

impl<'de, R> Deserializer<'de, R>where
R: XmlRead<'de>,

pub fn event_buffer_size(&mut self, limit: Option<NonZeroUsize>) -> &mut Self

Set the maximum number of events that could be skipped during deserialization of sequences.

If <element> contains more than specified nested elements, #text or CDATA nodes, then DeError::TooManyEvents will be returned during deserialization of sequence field (any type that uses deserialize_seq for the deserialization, for example, Vec<T>).

This method can be used to prevent a DoS attack and infinite memory consumption when parsing a very large XML to a sequence field.

It is strongly recommended to set limit to some value when you parse data from untrusted sources. You should choose a value that your typical XMLs can have between different elements that corresponds to the same sequence.

Examples

Let’s imagine, that we deserialize such structure:

struct List {
  item: Vec<()>,
}

The XML that we try to parse look like this:

<any-name>
  <item/>
  <!-- Bufferization starts at this point -->
  <another-item>
    <some-element>with text</some-element>
    <yet-another-element/>
  </another-item>
  <!-- Buffer will be emptied at this point; 7 events were buffered -->
  <item/>
  <!-- There is nothing to buffer, because elements follows each other -->
  <item/>
</any-name>

There, when we deserialize the item field, we need to buffer 7 events, before we can deserialize the second <item/>:

<another-item>

<some-element>

#text(with text)

</some-element>

<yet-another-element/> (virtual start event)

<yet-another-element/> (vitrual end event)

</another-item>

Note, that <yet-another-element/> internally represented as 2 events: one for the start tag and one for the end tag. In the future this can be eliminated, but for now we use auto-expanding feature of a reader, because this simplifies deserializer code.

impl<'de> Deserializer<'de, SliceReader<'de>>

pub fn from_str(s: &'de str) -> Self

Create new deserializer that will borrow data from the specified string

impl<'de, R> Deserializer<'de, IoReader<R>>where
R: BufRead,

pub fn from_reader(reader: R) -> Self

Create new deserializer that will copy data from the specified reader into internal buffer. If you already have a string use Self::from_str instead, because it will borrow instead of copy. If you have &[u8] which is known to represent UTF-8, you can decode it first before using from_str.