Struct quick_xml::events::attributes::Attribute

pub struct Attribute<'a> {
    pub key: &'a [u8],
    pub value: Cow<'a, [u8]>,
}

A struct representing a key/value XML attribute.

Field value stores raw bytes, possibly containing escape-sequences. Most users will likely want to access the value using one of the unescaped_value and unescape_and_decode_value functions.

Fields
key: &'a [u8]

The key to uniquely define the attribute.

If Attributes::with_checks is turned off, the key might not be unique.

value: Cow<'a, [u8]>

The raw value of the attribute.

Methods

impl<'a> Attribute<'a>

pub fn unescaped_value(&self) -> Result<Cow<[u8]>>

Returns the unescaped value.

This is normally the value you are interested in. Escape sequences such as &gt; are replaced with their unescaped equivalents such as >.

This will allocate if the value contains any escape sequences.

pub fn unescape_and_decode_value<B: BufRead>(
&self,
reader: &Reader<B>
) -> Result<String>

Returns the unescaped and decoded string value.

This allocates a String in all cases. For performance reasons it might be a better idea to instead use one of:

unescaped_value(), as it doesn't allocate when no escape sequences are used.

Reader::decode(), as it only allocates when the decoding can't be performed otherwise.