quick_xml::events::attributes

Struct Attribute

pub struct Attribute<'a> {
    pub key: QName<'a>,
    pub value: Cow<'a, [u8]>,
}

A struct representing a key/value XML attribute.

Field value stores raw bytes, possibly containing escape-sequences. Most users will likely want to access the value using one of the unescape_value and decode_and_unescape_value functions.

Fields
key: QName<'a>

The key to uniquely define the attribute.

If Attributes::with_checks is turned off, the key might not be unique.

value: Cow<'a, [u8]>

The raw value of the attribute.

Implementations

impl<'a> Attribute<'a>

pub fn unescape_value(&self) -> XmlResult<Cow<'a, str>>

Decodes using UTF-8 then unescapes the value.

This is normally the value you are interested in. Escape sequences such as &gt; are replaced with their unescaped equivalents such as >.

This will allocate if the value contains any escape sequences.

See also unescape_value_with()

This method is available only if encoding feature is not enabled.

pub fn unescape_value_with<'entity>( &self, resolve_entity: impl FnMut(&str) -> Option<&'entity str>, ) -> XmlResult<Cow<'a, str>>

Decodes using UTF-8 then unescapes the value, using custom entities.

This is normally the value you are interested in. Escape sequences such as &gt; are replaced with their unescaped equivalents such as >. A fallback resolver for additional custom entities can be provided via resolve_entity.

This will allocate if the value contains any escape sequences.

See also unescape_value()

This method is available only if encoding feature is not enabled.

pub fn decode_and_unescape_value( &self, decoder: Decoder, ) -> XmlResult<Cow<'a, str>>

Decodes then unescapes the value.

This will allocate if the value contains any escape sequences or in non-UTF-8 encoding.

pub fn decode_and_unescape_value_with<'entity>( &self, decoder: Decoder, resolve_entity: impl FnMut(&str) -> Option<&'entity str>, ) -> XmlResult<Cow<'a, str>>

Decodes then unescapes the value with custom entities.

This will allocate if the value contains any escape sequences or in non-UTF-8 encoding.

pub fn as_bool(&self) -> Option<bool>

If attribute value represents valid boolean values, returns Some, otherwise returns None.

The valid boolean representations are only "true", "false", "1", and "0".

Examples

use quick_xml::events::attributes::Attribute;

let attr = Attribute::from(("attr", "false"));
assert_eq!(attr.as_bool(), Some(false));

let attr = Attribute::from(("attr", "0"));
assert_eq!(attr.as_bool(), Some(false));

let attr = Attribute::from(("attr", "true"));
assert_eq!(attr.as_bool(), Some(true));

let attr = Attribute::from(("attr", "1"));
assert_eq!(attr.as_bool(), Some(true));

let attr = Attribute::from(("attr", "bot bool"));
assert_eq!(attr.as_bool(), None);