Struct quick_xml::events::attributes::Attribute

pub struct Attribute<'a> {
    pub key: &'a [u8],
    pub value: Cow<'a, [u8]>,
}

A struct representing a key/value XML attribute.

Field value stores raw bytes, possibly containing escape-sequences. Most users will likely want to access the value using one of the unescaped_value and unescape_and_decode_value functions.

Fields
key: &'a [u8]

The key to uniquely define the attribute.

If Attributes::with_checks is turned off, the key might not be unique.

value: Cow<'a, [u8]>

The raw value of the attribute.

Implementations

impl<'a> Attribute<'a>

pub fn unescaped_value(&self) -> Result<Cow<'_, [u8]>>

Returns the unescaped value.

This is normally the value you are interested in. Escape sequences such as &gt; are replaced with their unescaped equivalents such as >.

This will allocate if the value contains any escape sequences.

See also unescaped_value_with_custom_entities()

pub fn unescaped_value_with_custom_entities(
&self,
custom_entities: &HashMap<Vec<u8>, Vec<u8>>
) -> Result<Cow<'_, [u8]>>

Returns the unescaped value, using custom entities.

This is normally the value you are interested in. Escape sequences such as &gt; are replaced with their unescaped equivalents such as >. Additional entities can be provided in custom_entities.

This will allocate if the value contains any escape sequences.

See also unescaped_value()

Pre-condition

The keys and values of custom_entities, if any, must be valid UTF-8.

pub fn unescape_and_decode_value<B: BufRead>(
&self,
reader: &Reader<B>
) -> Result<String>

Decode then unescapes the value

This allocates a String in all cases. For performance reasons it might be a better idea to instead use one of:

Reader::decode(), as it only allocates when the decoding can’t be performed otherwise.

unescaped_value(), as it doesn’t allocate when no escape sequences are used.

pub fn unescape_and_decode_value_with_custom_entities<B: BufRead>(
&self,
reader: &Reader<B>,
custom_entities: &HashMap<Vec<u8>, Vec<u8>>
) -> Result<String>

Decode then unescapes the value with custom entities

This allocates a String in all cases. For performance reasons it might be a better idea to instead use one of:

Reader::decode(), as it only allocates when the decoding can’t be performed otherwise.

[unescaped_value()], as it doesn’t allocate when no escape sequences are used.

Pre-condition

The keys and values of custom_entities, if any, must be valid UTF-8.

pub fn unescape_and_decode_without_bom<B: BufRead>(
&self,
reader: &mut Reader<B>
) -> Result<String>

helper method to unescape then decode self using the reader encoding but without BOM (Byte order mark)

for performance reasons (could avoid allocating a String), it might be wiser to manually use

BytesText::unescaped()

Reader::decode(…)

pub fn unescape_and_decode_without_bom_with_custom_entities<B: BufRead>(
&self,
reader: &mut Reader<B>,
custom_entities: &HashMap<Vec<u8>, Vec<u8>>
) -> Result<String>

helper method to unescape then decode self using the reader encoding with custom entities but without BOM (Byte order mark)

for performance reasons (could avoid allocating a String), it might be wiser to manually use

BytesText::unescaped()

Reader::decode(…)

Pre-condition

The keys and values of custom_entities, if any, must be valid UTF-8.