Module quick_xml::name

Module for handling names according to the W3C Namespaces in XML 1.1 (Second Edition) specification

Structs

LocalName

A local (unqualified) name of an element or an attribute, i.e. a name without prefix.

Namespace

A namespace name that is declared in a xmlns[:prefix]="namespace name".

Prefix

A namespace prefix part of the qualified name of an element tag or an attribute: a prefix in <prefix:local-element-name> or prefix:local-attribute-name="attribute value".

QName

A qualified name of an element or an attribute, including an optional namespace prefix and a local name.

Enums

PrefixDeclaration

A namespace prefix declaration, xmlns or xmlns:<name>, as defined in XML Schema specification

ResolveResult

Result of prefix resolution which creates by NsReader::resolve_attribute, NsReader::resolve_element, NsReader::read_resolved_event and NsReader::read_resolved_event_into methods.