quick_xml

Module reader

Contains high-level interface for a pull-based XML parser.

Structs

BinaryStream

A direct stream to the underlying Readers reader which updates Reader::buffer_position() when read from it.

Config

A struct that holds a parser configuration.

NsReader

A low level encoding-agnostic XML event reader that performs namespace resolution.

Reader

A low level encoding-agnostic XML event reader.

Type Aliases

Span

Range of input in bytes, that corresponds to some piece of XML