Module quick_xml::reader

Contains high-level interface for a pull-based XML parser.

Modules

buffered_reader 🔒

This is an implementation of Reader for reading from a BufRead as underlying byte stream.

ns_reader 🔒

A reader that manages namespace declarations found in the input and able to resolve qualified names to expanded names.

slice_reader 🔒

This is an implementation of Reader for reading from a &[u8] as underlying byte stream. This implementation supports not using an intermediate buffer as the byte slice itself can be used to borrow from.

state 🔒

Macros

read_event_impl 🔒

read_to_end 🔒

Generalization of read_to_end method for buffered and borrowed readers

read_until_close 🔒

Read bytes up to the > and skip it. This method is expected to be called after seeing the < symbol and skipping it. Inspects the next (current) symbol and returns an appropriate Event:

Structs

BinaryStream

A direct stream to the underlying Readers reader which updates Reader::buffer_position() when read from it.

Config

A struct that holds a parser configuration.

NsReader

A low level encoding-agnostic XML event reader that performs namespace resolution.

Reader

A low level encoding-agnostic XML event reader.

Enums

BangType 🔒

Possible elements started with <!

ParseState 🔒

Possible reader states. The state transition diagram (true and false shows value of Config::expand_empty_elements option):

ReadTextResult 🔒

Result of an attempt to read XML textual data from the reader.

Traits

XmlSource 🔒

Represents an input for a reader that can return borrowed data.

Type Aliases

Span

Range of input in bytes, that corresponds to some piece of XML