quick_xml

Module se

Module to handle custom serde Serializer

Re-exports

pub use crate::errors::serialize::SeError;

Modules

content 🔒

Contains serializer for content of an XML element

element 🔒

Contains serializer for an XML element

key 🔒

simple_type 🔒

Contains Serde Serializer for XML simple types as defined in the XML Schema.

text 🔒

Contains serializer for a special &text field

Macros

forward 🔒

Implements serialization method by forwarding it to the serializer created by the helper method Serializer::ser.

write_primitive 🔒

Implements writing primitives to the underlying writer. Implementor must provide write_str(self, &str) -> Result<(), DeError> method

Structs

Serializer

A Serializer.

SimpleTypeSerializer

A serializer for a values representing XSD simple types, which used in:

XmlName 🔒

Helper struct to self-defense from errors

Enums

Indent 🔒

QuoteLevel

Defines which characters would be escaped in Text events and attribute values.

TextFormat

Defines the format for text content serialization

WriteResult

Classification of the type written by the serializer.

Functions

is_xml11_name_char 🔒

https://www.w3.org/TR/xml11/#NT-NameChar

is_xml11_name_start_char 🔒

Almost all characters can form a name. Citation from https://www.w3.org/TR/xml11/#sec-xml11:

to_string

Serialize struct into a String.

to_string_with_root

Serialize struct into a String using specified root tag name. root_tag should be valid XML name, otherwise error is returned.

to_utf8_io_writer

Serialize struct into a io::Writer restricted to utf-8 encoding.

to_writer

Serialize struct into a Writer.

to_writer_with_root

Serialize struct into a Writer using specified root tag name. root_tag should be valid XML name, otherwise error is returned.