quick_xml

Module se

Module to handle custom serde Serializer

Structs

Serializer

A Serializer

Enums

QuoteLevel

Defines which characters would be escaped in Text events and attribute values.

Functions

to_string

Serialize struct into a String.

to_string_with_root

Serialize struct into a String using specified root tag name. root_tag should be valid XML name, otherwise error is returned.

to_writer

Serialize struct into a Writer.

to_writer_with_root

Serialize struct into a Writer using specified root tag name. root_tag should be valid XML name, otherwise error is returned.